% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEI_B.R
\name{DEI_B}
\alias{DEI_B}
\title{Compute the p-values for the composite null hypothesis.}
\usage{
DEI_B(px, py)
}
\arguments{
\item{px}{a numeric vectors of p-values for the first individual hypothesis (e.g., the association between the exposure and mediator).}

\item{py}{a numeric vectors of p-values for the second individual hypothesis (e.g., the association between the mediator and outcome).}
}
\value{
The p-values for the composite null.
}
\description{
Compute the p-values for the composite null hypothesis.
}
\details{
Suppose that there are N items (e.g., mediators). Each item has two independent p-values.
Our method only calculates the composite null p-values for items whose maximum of the two individual p-values is less than 0.5. Otherwise, NA will be returned.
When the composite null p-values are large (e.g., > 0.1), they might not be accurate. Do not take these large p-values seriously (e.g., to draw a QQ plot).
}
\examples{
N = 100000  # number of items
set.seed(123)
### Generate the individual p-values from z scores. The first 100 items are under the composite alternative
x = rnorm(N); y = rnorm(N)
x[1:100] = x[1:100] + runif(100,min=4,max = 4)
y[1:100] = y[1:100] + runif(100,min=4,max = 4)
x[101:200]  = x[101:200] + runif(100,min=4,max = 4)
y[201:300]  = y[201:300] + runif(100,min=4,max = 4)
px = 2*pnorm(-abs(x)); py = 2*pnorm(-abs(y))
### Compute the p-values for the composite null.
pvals_compo = DEI_B(px,py)
### Apply the BH procedure to control FDR. Note there are many NA's in pvals_compo
pvals_compo[is.na(pvals_compo)] = 1
which(p.adjust(pvals_compo, method = "BH") < 0.05)
}
\references{
Liu, Y. (2025). A simple and powerful method for large-scale composite null hypothesis testing with applications in mediation analysis.  \emph{Biometrics}, 81(1), ujaf011.
}
\author{
Yaowu Liu
}
